+++ interface/index.js
@@ -1,5 +1,6 @@
 const callbackCommands = require('./callback_commands');
 const editQuestions = require('./edit_questions');
 const labels = require('./labels');
+const titles = require('./titles');
 
-module.exports = {callbackCommands, editQuestions, labels};
+module.exports = {callbackCommands, editQuestions, labels, titles};